+++ composer.json
@@ -6,10 +6,11 @@
     "license": "MIT",
     "require": {
         "php": "^8.1",
+        "ext-curl": "*",
         "ext-json": "*",
         "anlutro/l4-settings": "^1.0.0",
         "bensampo/laravel-enum": "^6.0",
-        "directorytree/ldaprecord": "^3.0",
+        "directorytree/ldaprecord-laravel": "^2.7",
         "doctrine/dbal": "^3.0",
         "goldspecdigital/laravel-eloquent-uuid": "^10",
         "guzzlehttp/guzzle": "^7.0.1",
@@ -19,8 +20,7 @@
         "laravel/tinker": "^2.0",
         "laravel/ui": "^4.0",
         "littleredbutton/bigbluebutton-api-php": "^4.2.0",
-        "maatwebsite/excel": "^3.1",
-      "ext-curl": "*"
+        "maatwebsite/excel": "^3.1"
     },
     "require-dev": {
         "ext-simplexml": "*",
@@ -47,7 +47,9 @@
     },
     "extra": {
         "laravel": {
-            "dont-discover": []
+            "dont-discover": [
+                "directorytree/ldaprecord-laravel"
+            ]
         }
     },
     "autoload": {
